break out company description
